Update build to Gradle 4.0
Fixes outdated web assumptions that caused the previous Gradle
version to fail.

NOTE: Somewhere in the 4.x series, Gradle stops recognizing the
default 'src' directory in this build, and I have not yet found
a workaround. In addition, 5.0 and later require a minimum of
Java 8, which this source has not been tested with. Upgrade at
your own risk.